+++ packages/core-js/internals/get-async-iterator.js
@@ -1,0 +1,15 @@
+var AsyncFromSyncIterator = require('../internals/async-from-sync-iterator');
+var aFunction = require('../internals/a-function');
+var anObject = require('../internals/an-object');
+var getIterator = require('../internals/get-iterator');
+var getMethod = require('../internals/get-method');
+var wellKnownSymbol = require('../internals/well-known-symbol');
+
+var ASYNC_ITERATOR = wellKnownSymbol('asyncIterator');
+
+module.exports = function (it) {
+  var method = getMethod(it[ASYNC_ITERATOR]);
+  if (method === undefined) {
+    return new AsyncFromSyncIterator(getIterator(it));
+  } return anObject(aFunction(method).call(it));
+};