+++ packages/core-js/modules/esnext.async-iterator.from.js
@@ -5,30 +5,35 @@ var aFunction = require('../internals/a-function');
 var anObject = require('../internals/an-object');
 var toObject = require('../internals/to-object');
 var createAsyncIteratorProxy = require('../internals/async-iterator-create-proxy');
-var getAsyncIteratorMethod = require('../internals/get-async-iterator-method');
+var getAsyncIterator = require('../internals/get-async-iterator');
+var getIteratorMethod = require('../internals/get-iterator-method');
+var getMethod = require('../internals/get-method');
+var wellKnownSymbol = require('../internals/well-known-symbol');
+var AsyncFromSyncIterator = require('../internals/async-from-sync-iterator');
+
+var ASYNC_ITERATOR = wellKnownSymbol('asyncIterator');
 
 var AsyncIterator = path.AsyncIterator;
 
 var AsyncIteratorProxy = createAsyncIteratorProxy(function (Promise, arg, hasArg) {
-  var step = anObject(this.next.apply(this.iterator, hasArg ? [arg] : []));
-  var done = step.done;
-  return Promise.resolve(step.value).then(function (value) {
-    return { done: done, value: value };
-  });
+  return anObject(this.next.apply(this.iterator, hasArg ? [arg] : []));
 }, true);
 
 $({ target: 'AsyncIterator', stat: true }, {
   from: function from(O) {
     var object = toObject(O);
-    var usingIterator = getAsyncIteratorMethod(object);
+    var usingIterator = getMethod(object[ASYNC_ITERATOR]);
     var iterator;
-    if (usingIterator != null) {
-      iterator = aFunction(usingIterator).call(object);
+    if (usingIterator !== undefined) {
+      iterator = getAsyncIterator(object);
       if (iterator instanceof AsyncIterator) return iterator;
-    } else {
-      iterator = object;
-    } return new AsyncIteratorProxy({
-      iterator: iterator
+    }
+    if (iterator === undefined) {
+      usingIterator = getIteratorMethod(object);
+      if (usingIterator !== undefined) return new AsyncFromSyncIterator(aFunction(usingIterator).call(object));
+    }
+    return new AsyncIteratorProxy({
+      iterator: iterator !== undefined ? iterator : object
     });
   }
 });